post3

+++ app/Http/Controllers/TasksController.php
@@ -15,15 +15,25 @@ class TasksController extends Controller
      */
     public function index()
     {
-        // メッセージ一覧を取得
-        $tasks = Task::all();         // 追加
-
-        // メッセージ一覧ビューでそれを表示
-        return view('tasks.index', [     // 追加
-            'tasks' => $tasks,        // 追加
-        ]);                                 // 追加
+        $data = [];
+        if (\Auth::check()) { // 認証済みの場合
+            // 認証済みユーザを取得
+            $user = \Auth::user();
+            // ユーザの投稿の一覧を作成日時の降順で取得
+            // （後のChapterで他ユーザの投稿も取得するように変更しますが、現時点ではこのユーザの投稿のみ取得します）
+            $tasks = $user->tasks()->orderBy('created_at', 'desc')->paginate(10);
+            $data = [
+//                'user' => $user,
+                'tasks' => $tasks,
+            ];
+            return view('tasks.index',$data);
+        }
+        
+        // dashboardビューでそれらを表示
+        return view('dashboard', $data);
     }
-
+    
+    
     /**
      * Show the form for creating a new resource.
      *
@@ -52,13 +62,14 @@ class TasksController extends Controller
             'content' => 'required|max:255',             
         ]);
         
-        $task = new task;
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        // 認証済みユーザ（閲覧者）の投稿として作成（リクエストされた値をもとに作成）
+        $request->user()->tasks()->create([
+             'status' => $request->status,
+            'content' => $request->content,
+        ]);
+        
+        // 前のURLへリダイレクトさせる
+          return redirect('/');
     }
 
     /**
@@ -88,11 +99,19 @@ class TasksController extends Controller
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
 
+
+        if (\Auth::id() === $task->user_id) {
+           return  view('tasks.edit', [
+                'task' => $task,  
+            ]);            
+        }
+        return redirect('/');
+        
         // メッセージ編集ビューでそれを表示
-        return view('tasks.edit', [
-            'task' => $task,
-           
-        ]);
+//        return view('tasks.edit', [
+//            'task' => $task,
+//           
+//        ]);
     }
 
     /**
@@ -111,11 +130,13 @@ class TasksController extends Controller
         
         // idの値でメッセージを検索して取得
         $task = Task::findOrFail($id);
-        // メッセージを更新
-        $task->status = $request->status;
-        $task->content = $request->content;
-        $task->save();
-
+    
+        if (\Auth::id() === $task->user_id) {        
+            // メッセージを更新
+            $task->status = $request->status;
+            $task->content = $request->content;
+            $task->save();
+        }
         // トップページへリダイレクトさせる
         return redirect('/');
     }
@@ -128,12 +149,21 @@ class TasksController extends Controller
      */
     public function destroy($id)
     {
-        // idの値でメッセージを検索して取得
-        $task = Task::findOrFail($id);
-        // メッセージを削除
-        $task->delete();
+        
+        
+        // idの値で投稿を検索して取得
+        $task = \App\Models\Task::findOrFail($id);
+        
+        // 認証済みユーザ（閲覧者）がその投稿の所有者である場合は投稿を削除
+        if (\Auth::id() === $task->user_id) {
+            $task->delete();
+            return back()
+                ->with('success','Delete Successful');
+        }
 
-        // トップページへリダイレクトさせる
-        return redirect('/');
+        // 前のURLへリダイレクトさせる
+        return back()
+            ->with('Delete Failed');
+
     }
 }
